keycloak_github_demo/requirements.txt
python-keycloak>=4.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
